move sub-package matrices according to unitest
