+++ quafu/elements/element_gates/unitary/decomposer.py
@@ -5,8 +5,8 @@ import numpy as np
 import scipy
 from numpy import ndarray
 
-from ..matrices import rz_mat, ry_mat, CXMatrix
-from ..matrices import mat_utils as mu
+from quafu.elements.matrices import rz_mat, ry_mat, CXMatrix
+from ...matrices import mat_utils as mu
 
 
 class UnitaryDecomposer(object):
